requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
openai==2.15.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
requests==2.32.5
rich==14.2.0
sniffio==1.3.1
textual==7.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.3
wcwidth==0.3.1
